Impleneted messaging system using xmppframework and jsqmessagingviewcontroller
Carlo want's us to use the actual database for the messaging, and the xmpp server just as notifications for when to recieve messages. I am currently not doing this, because since I'm already getting the data through xmpp It's easier, and less prone to bugs to just use it then. Currently the messaging works as a stand alone system. The idea is for it to be connected to the response video view, and then you can also send/recieve more videos (i think). This is not currently done, but shouldnt be too difficult to get them working together. Also note that I don't believe jsqmessagingviewcontroller supports media at this time. I had to use the XMPPframework and JSQMessageViewCOntroller to get everything working. THe xmpp framework had a lot of dependiencies, but they are all directly included in the project. Right now all of the xmpp functions work through the app delegate. This is so that it can recieve updates in a centralized place, and change settings when the app changes states. Carlo also thought it would be best if the messaging was handled appside, and the database was just used to send and recieve messages. Because of this, I created a singleton to handle the recieving of all messages. BFTMessageThreads recieves all messages, and keeps them in memory, and stores them locally. This is done so that even when The user is not on the messages page, they can still get updates. I haven't added any sort of notifications, but we certainly could. At some point the BFTMessageThreads class would best be handled using core data. Right now i am persisting data on the disk with nskeyedarchiver/unarchiver. THis keeps threads alive until the user specifically deletes them.